refactor(parser): split TypeScript parser adapter into focused modules
Extract typescript-parser-adapter.ts (1,012 lines) into:
- typescript-callable-parser.ts: code normalization, symbol keys, callable collection
- typescript-diff-matcher.ts: diff matching, grouping, summary generation
- typescript-parser-adapter.ts: thin facade (78 lines)

Prerequisite for H10 multi-language parser support.

Closes #171
